requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# docs/conf.py: 22
# python/gbpSID_dev/gbpSID_dev/project.py: 3
# python/gbpSID_dev/gbpSID_dev/scripts/update_gbpSID_docs.py: 4
# python/gbpSID_dev/setup.py: 8
GitPython == 2.1.8

# python/gbpSID_dev/gbpSID_dev/project.py: 4
PyYAML == 3.12

# docs/conf.py: 24
breathe == 4.7.3

# python/gbpSID_dev/gbpSID_dev/scripts/update_gbpSID_docs.py: 5
click == 6.7

# extern/Clara/scripts/embedTextFlow.py: 4
embed == 0.1.0

# docs/conf.py: 27
recommonmark == 0.4.0

# python/gbpSID_dev/setup.py: 3
setuptools == 38.5.1
